Move ColoursDataSet to separate file
